# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is a sample server Petstore server.  You can find out more about     Swagger at [http://swagger.io](http://swagger.io) or on [irc.freenode.net, #swagger](http://swagger.io/irc/).      For this sample, you can use the api key `special-key` to test the authorization     filters.",
    "version": "1.0.0",
    "title": "Swagger Petstore",
    "termsOfService": "http://swagger.io/terms/",
    "contact": {
      "email": "apiteam@swagger.io"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:1337",
  "basePath": "/",
...
  "schemes": [
    "https",
    "http"
  ],
  "paths": {
...
    "/packages/find": {
      "get": {
        "tags": [
          "packages"
        ],
        "summary": "Get all Packages",
        "description": "",
        "operationId": "getPackages",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Packages"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "Packages not found"
          }
        }
      }
    },
...
    "/connrenewal/find": {
      "get": {
        "tags": [
          "Connecion Renewal"
        ],
        "summary": "Get all Connecion Renewal",
        "description": "",
        "operationId": "getConnRenewal",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/ConnecionRenewal"
            }
          },
          "400": {
            "description": "Invalid id supplied"
          },
          "404": {
            "description": "ConnRenewal not found"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "Packages": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "package_name": {
          "type": "string"
        },
        "bandwidth": {
          "type": "string"
        },
        "data_limit": {
          "type": "string"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Packages Status"
        }
      },
      "xml": {
        "name": "Packages"
      }
    },
    "ConnecionRenewal": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "renewal_date": {
          "type": "string"
        },
        "connection_id": {
          "type": "number"
        },
        "status_id": {
          "type": "integer",
          "format": "int32",
          "description": "Connecion Renewal Status"
        }
      },
      "xml": {
        "name": "Connecion Renewal"
      }
    }
  },
...
}